Fix package structure and ranking endpoint stability
