+++ backend/src/main/java/com/mitosacademy/ranking/RankingApplication.java
@@ -1,6 +1,6 @@
 package com.mitosacademy.ranking;
 
-import config.RiotApiProperties;
+import com.mitosacademy.ranking.config.RiotApiProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -11,8 +11,8 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableScheduling
 public class RankingApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(RankingApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(RankingApplication.class, args);
+    }
 
 }
