+++ backend/src/main/java/com/mitosacademy/ranking/api/StatusController.java
@@ -1,4 +1,4 @@
-package api;
+package com.mitosacademy.ranking.api;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -11,4 +11,3 @@ public class StatusController {
         return "OK";
     }
 }
-
