+++ backend/src/main/java/com/mitosacademy/ranking/config/RiotApiProperties.java
@@ -1,4 +1,4 @@
-package config;
+package com.mitosacademy.ranking.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
@@ -27,4 +27,3 @@ public class RiotApiProperties {
         this.region = region;
     }
 }
-
