+++ backend/src/main/java/com/mitosacademy/ranking/domain/PlayerRanking.java
@@ -1,4 +1,4 @@
-package domain;
+package com.mitosacademy.ranking.domain;
 
 import java.time.OffsetDateTime;
